// src/app/model/sensorHumidityOfHourlyHumidity.ts
/**
 * Gimmie water applicaiton
 * Gardening application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { HourlyHumidity } from './hourlyHumidity';
import { Sensor } from './sensor';


export interface SensorHumidityOfHourlyHumidity { 
    humidityList?: Array<HourlyHumidity>;
    sensor?: Sensor;
}