// src/app/model/sensorHumidityOfDailyHumidity.ts
/**
 * Gimmie water applicaiton
 * Gardening application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DailyHumidity } from './dailyHumidity';
import { Sensor } from './sensor';


export interface SensorHumidityOfDailyHumidity { 
    humidityList?: Array<DailyHumidity>;
    sensor?: Sensor;
}